+++ newsy/templatetags/newsy_tags.py
@@ -1,6 +1,8 @@
+import operator
+
 from django import template
 
-from cms.templatetags.cms_tags import Placeholder
+from cms.templatetags.cms_tags import Placeholder, PluginsMedia
 
 from newsy.placeholders import render_newsy_placeholder
 
@@ -45,6 +47,10 @@ class NewsyPlaceholder(Placeholder):
         placeholder = _get_placeholder(context['current_page'], name)
         
         if placeholder:
+            request = context.get('request', None)
+            if request and hasattr(request, 'placeholder_media'):
+                request.placeholder_media = reduce(operator.add, [request.placeholder_media, placeholder.get_media(request, context)])
+        
             content = render_newsy_placeholder(placeholder, context, name)
         else:
             content = None
@@ -52,5 +58,24 @@ class NewsyPlaceholder(Placeholder):
         if not content and nodelist:
             return nodelist.render(context)
         return content
+
+class NewsyPluginsMedia(PluginsMedia):
+    name = 'newsy_plugins_media'
+    
+    def render_tag(self, context, page_lookup):
+        if not 'request' in context:
+            return ''
+        request = context['request']
+        from cms.plugins.utils import get_plugins_media
+        page = context.get('current_page', 'dummy')
+        if page == "dummy":
+            return ''
+        # make sure the plugin cache is filled
+        plugins_media = get_plugins_media(request, context, page)
+        
+        if plugins_media:
+            return plugins_media.render()
+        else:
+            return u''
 
 register.tag(NewsyPlaceholder)