+++ HISTORY.rst
@@ -5,6 +5,7 @@
 * Added tags field to NewsItem model for admin integration (GitHub issue #7)
 * Added tags field to NewsItem administration (GitHub issue #7)
 * Added south migration for the tags field (GitHub issue #7)
+* Added a news_plugins_media tag for rendering plugin media links (GitHub issue #8)
 
 0.1 (2011/07/15)
 ----------------